perf(rust/parser): shrink Token to u32 positions; +contiguity regression tests
PR2 of working/rust-parser-perf-rfc.md (Tier 1 item 4).

Token's start/end/line/column move from usize to u32, dropping the
struct from ~64 to ~48 bytes. tokenize() rejects sources whose byte
length doesn't fit in u32 (default max_size is 10MB; the cap is
defensive for callers who override it). Constructor signatures stay
usize for caller ergonomics; the narrowing happens once inside.

Consumers in parser_impl and parse_values cast `token.line`/`token.column`
back to usize at the boundary (ParseError::with_message and
OdinComment.line are unchanged).

Also adds 3 PR1 follow-up regression tests for the array contiguity
rewrite: index must start at 0, duplicate idx errors by default,
duplicate idx accepted under allow_duplicates with last-write-wins.

Bench (release, same fixtures), parse-* delta from PR1:
  parse-small  : 122,920 -> 126,814 ops/s
  parse-medium :  16,735 ->  16,822 ops/s
  parse-large  :     608 ->     763 ops/s (+25%)

Combined PR1+PR2 vs original baseline:
  parse-small  : 55,202 -> 126,814 (2.30x)
  parse-medium :  7,372 ->  16,822 (2.28x)
  parse-large  :    286 ->     763 (2.67x, beats RFC Target 700)

Tests: 4354 unit + 29 golden + integration all pass.

+++ odin-core/src/parser/parse_values.rs
@@ -48,37 +48,37 @@ pub fn parse_value<'a>(tokens: &[Token<'a>], pos: usize) -> Result<(OdinValue, u
                     // Bare strings are not allowed in ODIN
                     Err(ParseError::with_message(
                         ParseErrorCode::BareStringNotAllowed,
-                        token.line, token.column,
+                        token.line as usize, token.column as usize,
                         &format!("Unquoted string \"{}\" - use double quotes", token.value),
                     ))
                 }
             }
         }
         TokenType::NumberPrefix => {
-            let value = parse_number(&token.value, token.line, token.column)?;
+            let value = parse_number(&token.value, token.line as usize, token.column as usize)?;
             Ok((value, 1))
         }
         TokenType::IntegerPrefix => {
-            let value = parse_integer(&token.value, token.line, token.column)?;
+            let value = parse_integer(&token.value, token.line as usize, token.column as usize)?;
             Ok((value, 1))
         }
         TokenType::CurrencyPrefix => {
-            let value = parse_currency(&token.value, token.line, token.column)?;
+            let value = parse_currency(&token.value, token.line as usize, token.column as usize)?;
             Ok((value, 1))
         }
         TokenType::PercentPrefix => {
-            let value = parse_percent(&token.value, token.line, token.column)?;
+            let value = parse_percent(&token.value, token.line as usize, token.column as usize)?;
             Ok((value, 1))
         }
         TokenType::ReferencePrefix => {
             Ok((OdinValues::reference(&*token.value), 1))
         }
         TokenType::BinaryPrefix => {
-            let value = parse_binary(&token.value, token.line, token.column)?;
+            let value = parse_binary(&token.value, token.line as usize, token.column as usize)?;
             Ok((value, 1))
         }
         TokenType::DateLiteral => {
-            parse_date_value(&token.value, token.line, token.column)
+            parse_date_value(&token.value, token.line as usize, token.column as usize)
                 .map(|v| (v, 1))
         }
         TokenType::TimeLiteral => {
@@ -93,7 +93,7 @@ pub fn parse_value<'a>(tokens: &[Token<'a>], pos: usize) -> Result<(OdinValue, u
         TokenType::Path => {
             // Path tokens in value position can be temporal values
             if is_date_like(&token.value) {
-                if let Ok(val) = parse_date_value(&token.value, token.line, token.column) {
+                if let Ok(val) = parse_date_value(&token.value, token.line as usize, token.column as usize) {
                     return Ok((val, 1));
                 }
             }
@@ -110,7 +110,7 @@ pub fn parse_value<'a>(tokens: &[Token<'a>], pos: usize) -> Result<(OdinValue, u
             // Bare string — not allowed
             Err(ParseError::with_message(
                 ParseErrorCode::BareStringNotAllowed,
-                token.line, token.column,
+                token.line as usize, token.column as usize,
                 &format!("Unquoted string \"{}\" - use double quotes", token.value),
             ))
         }
@@ -160,7 +160,7 @@ pub fn parse_value<'a>(tokens: &[Token<'a>], pos: usize) -> Result<(OdinValue, u
         _ => {
             Err(ParseError::with_message(
                 ParseErrorCode::UnexpectedCharacter,
-                token.line, token.column,
+                token.line as usize, token.column as usize,
                 &format!("unexpected token type {:?} for value", token.token_type),
             ))
         }
